requirements/base.txt
pytz==2020.1
python-slugify==4.0.1
Pillow==7.2.0
argon2-cffi==20.1.0
whitenoise==5.2.0
redis==3.5.3
hiredis==1.1.0
celery==4.4.7
django-celery-beat==2.0.0
flower==0.9.5
sqlalchemy==1.3.19

# Django
# ------------------------------------------------------------------------------
django==3.0.9
django-environ==0.4.5
django-model-utils==4.0.0
django-allauth==0.42.0
django-crispy-forms==1.9.2
django-redis==4.12.1

# Django REST Framework
djangorestframework==3.11.1
django-rest-framework
django-cors-headers==3.4.0
